requirements.txt
apscheduler>=3.10.4
beautifulsoup4>=4.12.3
requests>=2.31.0
python-dotenv>=1.0.1
supabase>=2.4.0
google-generativeai>=0.7.0
pytz>=2023.3
feedparser>=6.0.10
python-dateutil>=2.8.2
